Touch: Add P8 compatibility
The P8 smartwatches use the CST716 or CST816S chips in various modes.
The CMake configuration variable DRIVER_TOUCH can be used to
set the factory-fused behaviour of these chips.

The device ID of these chips cannot be used for runtime detection,
because it does not give the hardware revision / firmware configuration.